Extract json_escape_string_without_quotes helper
